requirements.txt
# deep_research_agent requirements
# 全部免费，无需付费 API

# === Week 1: LangGraph 核心 ===
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0

# === LLM 客户端（按需装）===
langchain-google-genai>=2.0.0   # Gemini 免费 tier
langchain-groq>=0.2.0            # Groq 免费 tier
langchain-ollama>=0.2.0          # 本地 Ollama

# === Week 1: 工具 ===
ddgs>=6.0.0                      # 搜索（duckduckgo_search 已改名）
httpx>=0.27.0
beautifulsoup4>=4.12.0

# === Week 2: RAG + Memory ===
chromadb>=0.5.0                  # 向量库
sentence-transformers>=3.0.0     # BGE-M3 embedding
rank-bm25>=0.2.2                 # BM25 检索
langchain-huggingface>=0.1.0     # HuggingFace 集成

# === Week 2: 评估 ===
ragas>=0.2.0                     # RAG 三指标评估

# === 工具 ===
python-dotenv>=1.0.0